project/backend/requirements.txt
flask>=2.2.0
flask-cors>=3.0.0
python-dotenv>=1.0.0
werkzeug>=2.2.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.38
chromadb>=0.4.0
numpy>=1.20.0
transformers>=4.30.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
pypdf>=4.0.0